backend/requirements.txt
# Django Core
Django==5.0.10
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-extensions==3.2.3

# Django Real-time (Channels)
channels==4.2.0
channels-redis==4.2.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & Security
PyJWT==2.10.1
social-auth-core==4.5.4

# Social Media (X/Twitter)
tweepy==4.14.0
requests==2.31.0
oauthlib==3.2.2

# Reports
WeasyPrint==62.3
openpyxl==3.1.5
pandas==2.2.3

# ASGI Server (Channels)
daphne==4.1.2

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Development
python-decouple==3.8
python-dotenv==1.0.1

# Testing (Phase 1: Foundation)
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.25.3
factory-boy==3.3.1
